feat: export new plugin
