--- packages/solutions/app-tools/bin/modernv3.js
@@ -1,19 +1,0 @@
-#!/usr/bin/env node
-const {
-  INTERNAL_APP_TOOLS_PLUGINS,
-  INTERNAL_APP_TOOLS_RUNTIME_PLUGINS,
-} = require('@modern-js/utils');
-
-const { version } = require('../package.json');
-
-if (!process.env.MODERN_JS_VERSION) {
-  process.env.MODERN_JS_VERSION = version;
-}
-
-require('../dist/cjs/v3/run.js').run({
-  internalPlugins: {
-    cli: INTERNAL_APP_TOOLS_PLUGINS,
-    autoLoad: INTERNAL_APP_TOOLS_RUNTIME_PLUGINS,
-  },
-  initialLog: `Modern.js Framework v${version}`,
-});
