+++ packages/solutions/app-tools/src/new/compat/hooks.ts
