+++ packages/solutions/app-tools/src/new/compat/utils.ts
