+++ packages/solutions/app-tools/src/new/constants.ts
