+++ packages/solutions/app-tools/src/new/loadPlugins.ts
