+++ packages/solutions/app-tools/src/new/types/plugin.ts
